backend/requirements.txt
fastapi==0.115.7
pip-chill==1.0.3
psycopg2==2.9.10
pymongo==10.10.10.10
python-dotenv==1.0.1
uvicorn==0.34.0
